Add SPDY/3.1 support
The added feature is connection-level flow control.  The
FLOW_CONTROL_ERROR status code for GOAWAY frame is not specified in
the SPDY/3.1 spec, so we use PROTOCOL_ERROR instead in this commit.

The tools in src directory supports SPDY in clear, but the option to
specify SPDY protocol lacks SPDY/3.1 support.